cheats: fix assume not precompile (#594)
* cheats: fix assume not precompile

After the cancun hardfork, the point evaluation precompile was added
at `address(10)` which the `assumeNotPrecompile` implementation
did not take into account. Pectra is scheduled to include BLS
precompiles and likely more precompiles will be introduced in
other future hardforks. We might as well reserve the least significant
byte for Ethereum precompiles, as the address space is large and there
is no need to make calls from these addresses, and in return code
will not break on new EVM specs.

This was causing test failures in the Optimism test suite when
we updated the EVM version to cancun.

* cheats: update typo

Use the address range of precomiples that testnets use,
see https://github.com/eth-clients/holesky/blob/main/metadata/genesis.json