chore(`ci`): add CodeQL (#727)
<!--
Thank you for your Pull Request. Please provide a description above and
review
the requirements below.

Bug fixes and new features should include tests.

Contributors guide:
https://github.com/foundry-rs/foundry/blob/master/CONTRIBUTING.md

The contributors guide includes instructions for running rustfmt and
building the
documentation.
-->

<!-- ** Please select "Allow edits from maintainers" in the PR Options
** -->

## Motivation

<!--
Explain the context and why you're making that change. What is the
problem
you're trying to solve? In some cases there is not a problem and this
can be
thought of as being the motivation for your change.
-->

This PR introduces CodeQL code scanning initially just focused on Github
actions as it is fast to run.


https://docs.github.com/en/code-security/code-scanning/introduction-to-code-scanning/about-code-scanning-with-codeql

Results are reported privately in the `security` tab.

## Solution

<!--
Summarize the solution and provide any necessary context needed to
understand
the code change.
-->

This workflow was derived from the default workflow example Github
provides enhanced with concurrency cancel in progress, updated cron to
run daily and allow workflow dispatch. Trigger on cron, pull requests
and pushes to master.